edit the link button at informasiLayanan/index.blade.php

+++ app/Http/Controllers/SliderController.php
@@ -13,7 +13,7 @@ class SliderController extends Controller
     }
     
     public function create(){
-        // note langkah selanjutnya, tampilkan  dulu formnya, baru di lanjut ke logic kolom input order 'all order + 1, sebagai order baru' 
+        
         $totalOrdArr = Slider::pluck('order')->toArray();
         // dd($totalOrdArr);
         if(empty($totalOrdArr)){
@@ -22,14 +22,33 @@ class SliderController extends Controller
         $currOrd = max($totalOrdArr)+1;
         array_push($totalOrdArr, $currOrd);
         sort($totalOrdArr);
+        $link = Slider::pluck('link')->toArray();
+        // dd($link);
+        if(!count($totalOrdArr) < 3){
+            return redirect('/dashboard/slider');
+        }
         return view('dashboard.admin.cms.slider.create',[
-            'order' => $totalOrdArr
+            'order' => $totalOrdArr,
+            'link' => $link,
         ]);
     }
     public function store(Request $request){
+        $totalOrdArr = Slider::pluck('order')->toArray();
+        // dd($totalOrdArr);
+        if(empty($totalOrdArr)){
+            array_push($totalOrdArr, 0);
+        }
+        $currOrd = max($totalOrdArr)+1;
+        array_push($totalOrdArr, $currOrd);
+        sort($totalOrdArr);
+        
+        if(!count($totalOrdArr) < 3){
+            return redirect('/dashboard/slider');
+        }
         $request->validate([
             'img_url' => 'required|image|mimes:jpeg,jpg,png|max:2048',
             'caption' => 'required',
+            'link' => 'required',
             'order' => 'required',
             'is_active' => 'required',
         ]);
@@ -41,7 +60,6 @@ class SliderController extends Controller
         $allOrd = Slider::pluck('order')->toArray();
         for($i=0;$i<count($allOrd);$i++){
             if($allOrd[$i] == $request->order){
-                
                 Slider::where('order', $allOrd[$i])->update([
                     'order' => max($allOrd)+1
                 ]);
@@ -51,6 +69,7 @@ class SliderController extends Controller
             'img_url' => $namaFile,
             'caption' => $request->caption,
             'order' => $request->order,
+            'link' => $request->link,
             'is_active' => $request->is_active,
             'created_by' => 1,
             'lastupdate_by' => 1,
@@ -71,6 +90,7 @@ class SliderController extends Controller
         $request->validate([
             'img_url' => 'image|mimes:jpeg,jpg,png|max:2048',
             'caption' => 'required',
+            'link' => 'required',
             'order' => 'required',
             'is_active' => 'required',
         ]);
@@ -85,7 +105,7 @@ class SliderController extends Controller
         }else{
             $namaFile = $slider->img_url;
         }
-        // nanti jika sudah selesai logic dibawah ini, kita tanya ke gpt gimana cara cek data satu persatu, data request yang sesuai dengan data yang ada
+       
         $allOrd = Slider::pluck('order')->toArray();
         if($request->order !== $slider->order){
             for($i=0;$i<count($allOrd);$i++){
@@ -96,11 +116,17 @@ class SliderController extends Controller
             }
             }
         }
-        
+        if($request->link != $slider->link){
+            $otherSlider= Slider::where('link', $request->link)->first();
+            if($otherSlider){
+                $otherSlider->update(['link' => $slider->link]);
+            }
+        }
        Slider::where('id', $slider->id)->update([
             'img_url' => $namaFile,
             'caption' => $request->caption,
             'order' => $request->order,
+            'link' => $request->link,
             'is_active' => $request->is_active,
             'lastupdate_by' => 1,
        ]);
@@ -109,6 +135,17 @@ class SliderController extends Controller
     }
     
     public function destroy($id){
+        $totalOrdArr = Slider::pluck('order')->toArray();
+        // dd($totalOrdArr);
+        if(empty($totalOrdArr)){
+            array_push($totalOrdArr, 0);
+        }
+        $currOrd = max($totalOrdArr)+1;
+        array_push($totalOrdArr, $currOrd);
+        sort($totalOrdArr);
+        if($totalOrdArr == 1){
+            return redirect('/dashboard/slider');
+        }
         $slider = Slider::where('id', $id)->first();
         if(file_exists(public_path('uploads/slider/'.$slider->img_url))){
             unlink(public_path('/uploads/slider/'.$slider->img_url));
